+++ src/main/kotlin/bingsoochef/bingsoochef/toppping/presentation/req/CommentDto.kt
@@ -1,0 +1,5 @@
+package bingsoochef.bingsoochef.toppping.presentation.req
+
+data class CommentDto(
+    val commentContent: String
+)
