+++ src/main/kotlin/bingsoochef/bingsoochef/toppping/presentation/req/CreateToppingRequest.kt
@@ -1,4 +1,4 @@
-package bingsoochef.bingsoochef.toppping.req
+package bingsoochef.bingsoochef.toppping.presentation.req
 
 data class CreateToppingRequest (
     val bingsooId: Long,
