+++ src/main/kotlin/bingsoochef/bingsoochef/toppping/presentation/req/QuizDto.kt
@@ -1,10 +1,9 @@
-package bingsoochef.bingsoochef.toppping.res
+package bingsoochef.bingsoochef.toppping.presentation.req
 
 import bingsoochef.bingsoochef.toppping.domain.QuizType
 
 data class QuizDto(
-    val quizId: Long,
     val quizTitle: String,
     val quizType: QuizType,
-    val wrongCount: Short
+    val options: List<Pair<String, Boolean>>
 )
