+++ src/main/kotlin/bingsoochef/bingsoochef/toppping/presentation/req/ToppingDto.kt
@@ -1,4 +1,4 @@
-package bingsoochef.bingsoochef.toppping.req
+package bingsoochef.bingsoochef.toppping.presentation.req
 
 data class ToppingDto(
     val chefName: String,
