+++ src/main/kotlin/bingsoochef/bingsoochef/toppping/presentation/res/QuizDto.kt
@@ -1,9 +1,10 @@
-package bingsoochef.bingsoochef.toppping.req
+package bingsoochef.bingsoochef.toppping.presentation.res
 
 import bingsoochef.bingsoochef.toppping.domain.QuizType
 
 data class QuizDto(
+    val quizId: Long,
     val quizTitle: String,
     val quizType: QuizType,
-    val options: List<Pair<String, Boolean>>
+    val wrongCount: Short
 )
