+++ src/main/kotlin/bingsoochef/bingsoochef/toppping/presentation/res/TryResultResponse.kt
@@ -1,4 +1,4 @@
-package bingsoochef.bingsoochef.toppping.res
+package bingsoochef.bingsoochef.toppping.presentation.res
 
 data class TryResultResponse(
     val result: Boolean,
