--- src/main/kotlin/bingsoochef/bingsoochef/toppping/req/CommentDto.kt
@@ -1,5 +1,0 @@
-package bingsoochef.bingsoochef.toppping.req
-
-data class CommentDto(
-    val commentContent: String
-)
